+++ rows_test.go
@@ -8,11 +8,12 @@ import (
 	"testing"
 	"time"
 
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
 	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgconn"
 	"github.com/jackc/pgx/v5/pgxtest"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 type testRowScanner struct {
@@ -620,13 +621,14 @@ func TestRowToAddrOfStructPos(t *testing.T) {
 
 func TestRowToStructByName(t *testing.T) {
 	type person struct {
-		Last  string
-		First string
-		Age   int32
+		Last      string
+		First     string
+		Age       int32
+		AccountID string
 	}
 
 	defaultConnTestRunner.RunTest(context.Background(), t, func(ctx context.Context, t testing.TB, conn *pgx.Conn) {
-		rows, _ := conn.Query(ctx, `select 'John' as first, 'Smith' as last, n as age from generate_series(0, 9) n`)
+		rows, _ := conn.Query(ctx, `select 'John' as first, 'Smith' as last, n as age, 'd5e49d3f' as account_id from generate_series(0, 9) n`)
 		slice, err := pgx.CollectRows(rows, pgx.RowToStructByName[person])
 		assert.NoError(t, err)
 
@@ -635,6 +637,7 @@ func TestRowToStructByName(t *testing.T) {
 			assert.Equal(t, "Smith", slice[i].Last)
 			assert.Equal(t, "John", slice[i].First)
 			assert.EqualValues(t, i, slice[i].Age)
+			assert.Equal(t, "d5e49d3f", slice[i].AccountID)
 		}
 
 		// check missing fields in a returned row
@@ -643,7 +646,7 @@ func TestRowToStructByName(t *testing.T) {
 		assert.ErrorContains(t, err, "cannot find field First in returned row")
 
 		// check missing field in a destination struct
-		rows, _ = conn.Query(ctx, `select 'John' as first, 'Smith' as last, n as age, null as ignore from generate_series(0, 9) n`)
+		rows, _ = conn.Query(ctx, `select 'John' as first, 'Smith' as last, n as age, 'd5e49d3f' as account_id, null as ignore from generate_series(0, 9) n`)
 		_, err = pgx.CollectRows(rows, pgx.RowToAddrOfStructByName[person])
 		assert.ErrorContains(t, err, "struct doesn't have corresponding row field ignore")
 	})